feat : MovieModal design 부분 component화
