+++ core/design/build.gradle.kts
@@ -22,4 +22,6 @@ dependencies {
     api(libs.androidx.compose.material3)
     debugApi(libs.androidx.compose.ui.tooling)
     api(libs.androidx.compose.ui.tooling.preview)
+
+    implementation(libs.a914.gowtham.compose.ratingbar)
 }
